requirements.txt
pyshacl>=0.25
pyyaml>=6.0
rdflib>=7.0.0
